Move Status to core.game
