+++ modules/game/src/main/Env.scala
@@ -72,13 +72,11 @@ final class Env(
 
   lazy val importer = wire[lila.game.importer.Importer]
 
-  val statusText: lila.core.game.StatusText = StatusText.apply
-
   lazy val api: lila.core.game.GameApi = new:
     export gameRepo.{ incBookmarks, getSourceAndUserIds }
     export cached.nbPlaying
     export GameExt.{ computeMoveTimes, analysable }
-    export StatusText.apply as statusText
+    export lila.core.game.StatusText.apply as statusText
     export AnonCookie.json as anonCookieJson
 
   lazy val newPlayer: lila.core.game.NewPlayer = new:
